pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "linkedin-poster"
version = "0.1.0"
description = "CLI tool for auto-generating and posting LinkedIn content showcasing GitHub POC projects"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "respx>=0.20.0",
]

[project.scripts]
linkedin-poster = "linkedin_poster.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
